backend/pyproject.toml
[project]
name = "ctf-platform-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.0",
    "docker>=7.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
